Refactor getElasticConfig and getElasticSearchClient to a trait in order to remove duplication. Support elasticsearch 2.0 create from Builder.
